+++ {{cookiecutter.repo_name}}/resources/docs/conf.py
@@ -2,9 +2,6 @@
 {% if cookiecutter.sphinx_theme == 'sphinx-rtd-theme' -%}
 import os
 {% endif -%}
-{%- if cookiecutter.setup_py_uses_setuptools_scm == 'yes' -%}
-import traceback
-{% endif -%}
 {%- if cookiecutter.sphinx_theme != 'sphinx-rtd-theme' -%}
 import {{ cookiecutter.sphinx_theme|replace('-', '_') }}
 {% endif %}
@@ -23,21 +20,11 @@ extensions = [
 source_suffix = ['.rst']
 
 master_doc = 'index'
-project = {{ '{0!r}'.format(cookiecutter.project_name) }}
-year = '{% if cookiecutter.year_from == cookiecutter.year_to %}{{ cookiecutter.year_from }}{% else %}{{ cookiecutter.year_from }}-{{ cookiecutter.year_to }}{% endif %}'
+project = {{ '{0!r}'.format(cookiecutter.repo_name) }}
+year = '{% if cookiecutter.year == cookiecutter.year_to %}{{ cookiecutter.year }}{% else %}{{ cookiecutter.year }}-{{ cookiecutter.year_to }}{% endif %}'
 author = {{ '{0!r}'.format(cookiecutter.full_name) }}
 copyright = '{0}, {1}'.format(year, author)
-{%- if cookiecutter.setup_py_uses_setuptools_scm == 'yes' %}
-try:
-    from pkg_resources import get_distribution
-
-    version = release = get_distribution('{{ cookiecutter.package_name }}').version
-except Exception:
-    traceback.print_exc()
-    version = release = {{ '{0!r}'.format(cookiecutter.version) }}
-{%- else %}
 version = release = {{ '{0!r}'.format(cookiecutter.version) }}
-{%- endif %}
 
 pygments_style = 'trac'
 templates_path = ['.']