+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.package_name}}/cli.py
@@ -19,44 +19,15 @@ problems: the code will get executed twice:
 
 Also see (1) from http://click.pocoo.org/5/setuptools/#setuptools-integration
 """
-{%- if cookiecutter.command_line_interface == 'click' %}
-import click
-{%- elif cookiecutter.command_line_interface == 'argparse' %}
+
 import argparse
-{%- else %}
-from typing import Literal
-import sys
-{%- endif %}
-{%- if cookiecutter.command_line_interface == 'click' %}
 
-
-@click.command()
-@click.argument('names', nargs=-1)
-def main(names) -> None:
-    click.echo(repr(names))
-{%- elif cookiecutter.command_line_interface == 'argparse' %}
-
-parser = argparse.ArgumentParser(description='Command description.')
-parser.add_argument('names', metavar='NAME', nargs=argparse.ZERO_OR_MORE,
-                    help="A name of something.")
+parser = argparse.ArgumentParser(description="Command description.")
+parser.add_argument(
+    "names", metavar="NAME", nargs=argparse.ZERO_OR_MORE, help="A name of something."
+)
 
 
 def main(args=None) -> None:
     args = parser.parse_args(args=args)
     print(args.names)
-{%- else %}
-
-
-def main(argv=sys.argv) -> Literal[0]:
-    """
-    Args:
-        argv (list): List of arguments
-
-    Returns:
-        int: A return code
-
-    Does stuff.
-    """
-    print(argv)
-    return 0
-{%- endif %}