+++ CHANGELOG.rst
@@ -47,7 +47,7 @@ v1.15 (2019-05-25)
 ------------------
 
 * Updated various URLS.
-* Changed ``years`` to ``year_from`` and ``year_to`` to have less fuss when regenerating a project.
+* Changed ``years`` to ``year`` and ``year_to`` to have less fuss when regenerating a project.
 * Fixed various configuration issues (broken envs, coverage configuration, travis configuration etc).
 * Added project urls and python version requirements in ``setup.py``.